+++ R/plot_variable_dropout.R
@@ -1,9 +1,10 @@
-#' Plots Global Model Explanations (Variable Drop-out)
+#' Plots Global Model Explanations (Variable Importance)
 #'
 #' Function \code{plot.variable_dropout_explainer} plots dropouts for variables used in the model.
 #' It uses output from \code{variable_dropout} function that corresponds to permutation based measure of variable importance.
+#' Variables are sorted in the same order in all panels. The order depends on the average drop out loss. In different panels variable contributions may not look like sorted if variable importance is different in different in different mdoels.
 #'
-#' @param x a variable fropout exlainer produced with the 'variable_dropout' function
+#' @param x a variable dropout exlainer produced with the 'variable_dropout' function
 #' @param ... other explainers that shall be plotted together
 #' @param max_vars maximum number of variables that shall be presented for for each model
 #'
@@ -18,14 +19,14 @@
 #' library("randomForest")
 #' HR_rf_model <- randomForest(left~., data = breakDown::HR_data, ntree = 100)
 #' explainer_rf  <- explain(HR_rf_model, data = HR_data, y = HR_data$left)
-#' vd_rf <- variable_dropout(explainer_rf, type = "raw")
+#' vd_rf <- variable_importance(explainer_rf, type = "raw")
 #' vd_rf
 #' plot(vd_rf)
 #'
 #' HR_glm_model <- glm(left~., data = breakDown::HR_data, family = "binomial")
 #' explainer_glm <- explain(HR_glm_model, data = HR_data, y = HR_data$left)
 #' logit <- function(x) exp(x)/(1+exp(x))
-#' vd_glm <- variable_dropout(explainer_glm, type = "raw",
+#' vd_glm <- variable_importance(explainer_glm, type = "raw",
 #'                         loss_function = function(observed, predicted)
 #'                                    sum((observed - logit(predicted))^2))
 #' vd_glm
@@ -39,7 +40,7 @@
 #' HR_xgb_model <- xgb.train(param, data_train, nrounds = 50)
 #' explainer_xgb <- explain(HR_xgb_model, data = model_martix_train,
 #'                                     y = HR_data$left, label = "xgboost")
-#' vd_xgb <- variable_dropout(explainer_xgb, type = "raw")
+#' vd_xgb <- variable_importance(explainer_xgb, type = "raw")
 #' vd_xgb
 #' plot(vd_xgb)
 #'
@@ -47,15 +48,15 @@
 #'
 #' # NOTE:
 #' # if you like to have all importances hooked to 0, you can do this as well
-#' vd_rf <- variable_dropout(explainer_rf, type = "difference")
-#' vd_glm <- variable_dropout(explainer_glm, type = "difference",
+#' vd_rf <- variable_importance(explainer_rf, type = "difference")
+#' vd_glm <- variable_importance(explainer_glm, type = "difference",
 #'                         loss_function = function(observed, predicted)
 #'                                    sum((observed - logit(predicted))^2))
-#' vd_xgb <- variable_dropout(explainer_xgb, type = "difference")
+#' vd_xgb <- variable_importance(explainer_xgb, type = "difference")
 #' plot(vd_rf, vd_glm, vd_xgb)
 #' }
 #'
-plot.variable_dropout_explainer <- function(x, ..., max_vars = 10) {
+plot.variable_importance_explainer <- function(x, ..., max_vars = 10) {
   dfl <- c(list(x), list(...))
 
   # combine all explainers in a single frame