+++ R/single_prediction.R
@@ -7,9 +7,11 @@
 #' @return An object of the class 'single_prediction_explainer'.
 #' It's a data frame with calculated average response.
 #'
+#' @aliases single_prediction
 #' @export
 #' @import breakDown
 #' @examples
+#' \dontrun{
 #' library("breakDown")
 #' new.wine <- data.frame(citric.acid = 0.35,
 #'      sulphates = 0.6,
@@ -19,14 +21,13 @@
 #'
 #' wine_lm_model4 <- lm(quality ~ pH + residual.sugar + sulphates + alcohol, data = wine)
 #' wine_lm_explainer4 <- explain(wine_lm_model4, data = wine, label = "model_4v")
-#' wine_lm_predict4 <- single_prediction(wine_lm_explainer4, observation = new.wine)
+#' wine_lm_predict4 <- prediction_breakdown(wine_lm_explainer4, observation = new.wine)
 #' wine_lm_predict4
 #'
-#' \dontrun{
 #' library("randomForest")
 #' wine_rf_model4 <- randomForest(quality ~ pH + residual.sugar + sulphates + alcohol, data = wine)
 #' wine_rf_explainer4 <- explain(wine_rf_model4, data = wine, label = "model_rf")
-#' wine_rf_predict4 <- single_prediction(wine_rf_explainer4, observation = new.wine)
+#' wine_rf_predict4 <- prediction_breakdown(wine_rf_explainer4, observation = new.wine)
 #' wine_rf_predict4
 #'
 #' library("gbm")
@@ -39,27 +40,29 @@
 #'              n.minobsinnode = 10,
 #'              verbose = FALSE)
 #'  # make an explainer for the model
-#'  explainer_gbm <- explain(model, data = wine)
+#'  explainer_gbm <- explain(model, data = wine, predict_function =
+#'          function(model, x) predict(model, x, n.trees = 1000))
 #'  # create a new observation
-#'  exp_sgn <- single_prediction(explainer_gbm, observation = new.wine,
-#'               n.trees = 1000)
+#'  exp_sgn <- prediction_breakdown(explainer_gbm, observation = new.wine)
 #'  exp_sgn
 #'  }
 #'
-single_prediction <- function(explainer, observation, ...) {
-  if (!("explainer" %in% class(explainer))) stop("The single_prediction() function requires an object created with explain() function.")
-  if (is.null(explainer$data)) stop("The single_prediction() function requires explainers created with specified 'data' parameter.")
+prediction_breakdown <- function(explainer, observation, ...) {
+  if (!("explainer" %in% class(explainer))) stop("The prediction_breakdown() function requires an object created with explain() function.")
+  if (is.null(explainer$data)) stop("The prediction_breakdown() function requires explainers created with specified 'data' parameter.")
 
   # breakDown
   res <- broken(explainer$model,
                 new_observation = observation,
                 data = explainer$data,
+                predict.function = explainer$predict_function,
                 baseline = "Intercept", ...)
   res$label <- rep(explainer$label, length(res$variable))
 
-  class(res) <- c("single_prediction_explainer", "data.frame")
+  class(res) <- c("prediction_breakdown_explainer", "data.frame")
   res
 
 }
 
-
+#' @export
+single_prediction <- prediction_breakdown